:app — split Settings into a root list + seven sub-pages
The Settings screen was one ~1,430-line composable cramming every
section into a single scroll. Split it into a root list of seven
tappable rows, each drilling into a focused sub-page:

  Schedule (time, days, delivery mode)
  Wardrobe (rules)
  Voice    (TTS engine, voice, locale, test)
  Units    (temperature, distance)
  API keys (Gemini, OpenAI, ElevenLabs)
  Data sources (location, calendar)
  About    (version, privacy, links; Debug in debug builds)

Navigation reuses MainActivity's rememberSaveable / BackHandler
pattern — Compose Navigation is still overkill at one level deep.
Delivery Mode moves under Schedule per future "different delivery
per schedule" plans. Voice locale is promoted out of the Device-
engine branch so it's reachable from any engine choice (resolves
the existing TODO; locale also drives the OpenAI default voice).

No behaviour or state changes — same SettingsState, same VM, same
strings (plus seven short row/title labels).
