+++ app/src/main/kotlin/app/adaptweather/ui/settings/ScheduleSettings.kt
@@ -1,0 +1,172 @@
+package app.adaptweather.ui.settings
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.FilterChip
+import androidx.compose.material3.FilterChipDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TimePicker
+import androidx.compose.material3.rememberTimePickerState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import app.adaptweather.R
+import app.adaptweather.core.domain.model.DeliveryMode
+import java.time.DayOfWeek
+import java.time.LocalTime
+import java.time.format.DateTimeFormatter
+import java.time.format.TextStyle
+import java.util.Locale
+
+@Composable
+internal fun ScheduleContent(
+    time: LocalTime,
+    days: Set<DayOfWeek>,
+    deliveryMode: DeliveryMode,
+    padding: PaddingValues,
+    onSetSchedule: (LocalTime, Set<DayOfWeek>) -> Unit,
+    onSetDeliveryMode: (DeliveryMode) -> Unit,
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(padding)
+            .verticalScroll(rememberScrollState())
+            .padding(horizontal = 16.dp, vertical = 8.dp),
+        verticalArrangement = Arrangement.spacedBy(12.dp),
+    ) {
+        ScheduleCard(time, days, onSetSchedule)
+        DeliveryModeCard(deliveryMode, onSetDeliveryMode)
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+private fun ScheduleCard(
+    time: LocalTime,
+    days: Set<DayOfWeek>,
+    onChange: (LocalTime, Set<DayOfWeek>) -> Unit,
+) {
+    var pickerOpen by remember { mutableStateOf(false) }
+
+    SectionCard(title = stringResource(R.string.settings_schedule_title)) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically,
+        ) {
+            Text(
+                text = stringResource(R.string.settings_schedule_time_label),
+                style = MaterialTheme.typography.bodyMedium,
+                modifier = Modifier.padding(end = 12.dp),
+            )
+            OutlinedButton(onClick = { pickerOpen = true }) {
+                Text(text = TIME_FORMAT.format(time))
+            }
+        }
+
+        Text(
+            text = stringResource(R.string.settings_schedule_days_label),
+            style = MaterialTheme.typography.bodyMedium,
+        )
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.spacedBy(4.dp),
+        ) {
+            DayOfWeek.entries.forEach { dow ->
+                val selected = dow in days
+                FilterChip(
+                    selected = selected,
+                    onClick = {
+                        val next = if (selected) days - dow else days + dow
+                        if (next.isNotEmpty()) onChange(time, next)
+                    },
+                    label = {
+                        Text(text = dow.getDisplayName(TextStyle.SHORT, Locale.getDefault()))
+                    },
+                    colors = FilterChipDefaults.filterChipColors(),
+                )
+            }
+        }
+    }
+
+    if (pickerOpen) {
+        TimePickerDialog(
+            initial = time,
+            onDismiss = { pickerOpen = false },
+            onConfirm = { newTime ->
+                pickerOpen = false
+                onChange(newTime, days)
+            },
+        )
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+private fun TimePickerDialog(
+    initial: LocalTime,
+    onDismiss: () -> Unit,
+    onConfirm: (LocalTime) -> Unit,
+) {
+    val state = rememberTimePickerState(
+        initialHour = initial.hour,
+        initialMinute = initial.minute,
+        is24Hour = true,
+    )
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        confirmButton = {
+            TextButton(onClick = { onConfirm(LocalTime.of(state.hour, state.minute)) }) {
+                Text(stringResource(android.R.string.ok))
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDismiss) {
+                Text(stringResource(android.R.string.cancel))
+            }
+        },
+        text = { TimePicker(state = state) },
+    )
+}
+
+@Composable
+private fun DeliveryModeCard(
+    selected: DeliveryMode,
+    onSelect: (DeliveryMode) -> Unit,
+) {
+    SectionCard(title = stringResource(R.string.settings_delivery_title)) {
+        DeliveryMode.entries.forEach { mode ->
+            RadioRow(
+                label = stringResource(deliveryModeLabel(mode)),
+                selected = mode == selected,
+                onSelect = { onSelect(mode) },
+            )
+        }
+    }
+}
+
+private fun deliveryModeLabel(mode: DeliveryMode): Int = when (mode) {
+    DeliveryMode.NOTIFICATION_ONLY -> R.string.settings_delivery_notification_only
+    DeliveryMode.TTS_ONLY -> R.string.settings_delivery_tts_only
+    DeliveryMode.NOTIFICATION_AND_TTS -> R.string.settings_delivery_notification_and_tts
+}
+
+private val TIME_FORMAT: DateTimeFormatter = DateTimeFormatter.ofPattern("HH:mm")
